Refector: 임시 컨텍스트 삭제
