--- src/contexts/issueContextTempForIssueList.tsx
@@ -1,57 +1,0 @@
-import { ReactNode, createContext, useRef, useState } from 'react';
-import { getIssues } from '../api/issueApi';
-import { Issues } from '../types/Issue';
-import { ORGANIZATION, REPO } from '../utils/constants';
-
-interface GlobalContextPops {
-  issues: Issues;
-  isLoading: boolean;
-  isError: boolean;
-  fetchIssues: () => Promise<void>;
-}
-
-const GlobalContext: GlobalContextPops = {
-  issues: [],
-  isLoading: false,
-  isError: false,
-  fetchIssues: () => {
-    throw new Error();
-  },
-};
-
-export const IssueContextProviderTemp = createContext(GlobalContext);
-
-export function IssueContextProvider({ children }: { children: ReactNode }) {
-  const [issues, setIssues] = useState<Issues>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isError, setIsError] = useState(false);
-
-  const pageRef = useRef(1);
-  const isEndRef = useRef(false);
-
-  const fetchIssues = async () => {
-    if (isEndRef.current) return;
-    setIsLoading(true);
-    try {
-      const newIssues = await getIssues(ORGANIZATION, REPO, pageRef.current);
-      if (newIssues.length === 0) {
-        isEndRef.current = true;
-        return;
-      }
-      pageRef.current = pageRef.current + 1;
-      setIssues(prevIssues => [...prevIssues, ...newIssues]);
-    } catch (error) {
-      setIsError(true);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <IssueContextProviderTemp.Provider
-      value={{ issues, isLoading, fetchIssues, isError }}
-    >
-      {children}
-    </IssueContextProviderTemp.Provider>
-  );
-}
