+++ src/page/Home.tsx
@@ -1,6 +1,6 @@
 import useInfinityScroll from 'hooks/useInfinityScroll';
 import { useContext } from 'react';
-import { IssueContextProviderTemp } from 'contexts/issueContextTempForIssueList';
+import { IssueContext } from 'contexts/IssueContext';
 import { styled } from 'styled-components';
 import AdvertisementItem from 'components/Advertisement';
 import IssuesList from 'components/IssuesList';
@@ -8,9 +8,7 @@ import Error from 'elements/Error';
 import Loading from 'elements/Loading';
 
 const Home: React.FC = () => {
-  const { issues, fetchIssues, isLoading, isError } = useContext(
-    IssueContextProviderTemp
-  );
+  const { issues, fetchIssues, isLoading, isError } = useContext(IssueContext);
 
   const observeRef = useInfinityScroll({ callback: getIssuesCallback });
